+++ CHANGES.md
@@ -11,6 +11,8 @@
 - Fixed automatic conversion for the action-group
 - Reverted the layout-flow component, auto-generation broke it 
 - make tests run with `python -m pytest tests/`
+- extend parser to process shorthand boolean attributes
+- add error handling around shorthand boolean attributes
 
 # [0.1] - 2025-11-04
 - First tagged version